Add check authentication security in Ajax files and logout

+++ ajax/deleteuser.php
@@ -2,7 +2,11 @@
 
 require $_SERVER['DOCUMENT_ROOT'] . '/include/autoloader.php';
 
-$user = new User();
-$user->delete();
-
+if ($auth->getSessIdNum() != 0) {
+    $user->delete();
+} else {
+    header("HTTP/1.1 401 Unauthorized");
+    echo "Error: 401 Unauthorized";
+    exit;
+}
 ?>
